requirements.txt
# ============================================
# NCERT AI Tutor - Requirements
# Django-based AI Tutor Application
# ============================================

# Core Django
Django>=5.2.0
djangorestframework>=3.14.0
django-cors-headers>=4.0.0

# Database (Optional: Use PostgreSQL in production)
psycopg2-binary>=2.9.0

# AI & ML - Local LLM
ollama>=0.1.0

# Core AI and ML libraries (from original)
transformers==4.41.2
accelerate
bitsandbytes
torch>=2.0.0

# API Fallbacks (Optional)
openai>=1.0.0
anthropic>=0.7.0

# RAG and Data Processing
sentence-transformers
chromadb
PyMuPDF

# Math PDF processing
nougat-ocr==0.1.17
albumentations==1.3.1

# Web Scraping
beautifulsoup4>=4.12.0
requests>=2.31.0
lxml>=4.9.0

# Multimodal Support
openai-whisper>=20231117  # Speech-to-text
gtts>=2.4.0  # Text-to-speech
matplotlib>=3.7.0  # Diagram generation
numpy>=1.24.0
Pillow>=10.0.0

# Utilities
python-dotenv>=1.0.0
hashlib3>=0.0.1

# Caching & Background Tasks (Optional)
celery>=5.3.0
redis>=5.0.0

# Development
pytest>=7.4.0
pytest-django>=4.5.0
black>=23.0.0
flake8>=6.0.0

# Production
gunicorn>=21.2.0
whitenoise>=6.5.0  # Static file serving
daphne>=4.0.0  # ASGI server